Move channel creation with dispersy function to more general class
